refactor(lightbox): improve readability of box removal logic
