+++ apps/extension/entrypoints/fiber-reader.ts
@@ -1,5 +1,5 @@
-import { extractComponentAncestry } from '../utils/fiber-component';
-import { FIBER_REQUEST, FIBER_RESPONSE, FIBER_TARGET_ATTR } from '../utils/fiber-protocol';
+import { extractComponentAncestry } from '../utils/dom/fiber-component';
+import { FIBER_REQUEST, FIBER_RESPONSE, FIBER_TARGET_ATTR } from '../utils/messaging/fiber-protocol';
 
 // MAIN world에서 실행 — isolated content script는 fiber expando를 못 읽으므로,
 // 마킹된 노드의 컴포넌트 계보를 여기서 읽어 postMessage로 돌려준다.
