+++ apps/extension/entrypoints/inspector.content/fab.tsx
