+++ apps/extension/entrypoints/inspector.content/index.ts
@@ -1,15 +1,15 @@
 import { createElement } from 'react';
 import { createRoot } from 'react-dom/client';
 
-import { FIBER_REQUEST, FIBER_RESPONSE, FIBER_TARGET_ATTR } from '../../utils/fiber-protocol';
-import type { FiberResponse } from '../../utils/fiber-protocol';
+import { FIBER_REQUEST, FIBER_RESPONSE, FIBER_TARGET_ATTR } from '../../utils/messaging/fiber-protocol';
+import type { FiberResponse } from '../../utils/messaging/fiber-protocol';
 import { onMessage, sendMessage } from '../../utils/messaging';
-import { buildSelector } from '../../utils/selector';
-import { addItem } from '../../utils/session-store';
-import type { CapturedRect } from '../../utils/session-store';
-import { extractStyle } from '../../utils/style-extract';
-import { InspectorUi } from './InspectorUi';
-import type { InspectorUiHandle } from './InspectorUi';
+import { buildSelector } from '../../utils/dom/selector';
+import { addItem } from '../../utils/data/session-store';
+import type { CapturedRect } from '../../utils/data/session-store';
+import { extractStyle } from '../../utils/dom/style-extract';
+import { InspectorUi } from './inspector-ui';
+import type { InspectorUiHandle } from './inspector-ui';
 import { showLightbox } from './lightbox';
 import { createOverlay } from './overlay';
