+++ apps/extension/entrypoints/popup/useQaSession.ts
@@ -1,7 +1,7 @@
 import { useCallback, useEffect, useState } from 'react';
 
-import { clearImages } from '../../utils/image-store';
-import { clearItems, getItems, watchItems } from '../../utils/session-store';
+import { clearImages } from '../../utils/data/image-store';
+import { clearItems, getItems, watchItems } from '../../utils/data/session-store';
 
 export interface UseQaSession {
     count: number;
