+++ apps/extension/utils/data/image-store.test.ts
