+++ apps/extension/utils/data/image-store.ts
