+++ apps/extension/utils/data/session-store.ts
