+++ apps/extension/utils/dom/fiber-component.test.ts
