+++ apps/extension/utils/dom/fiber-component.ts
