+++ apps/extension/utils/dom/selector.ts
